🗺️ Atlas: [architectural change] Refactor building.rs Blob into building module with configuration submodule
Extracted the massive "God Function" `spawn_building` and its 15 associated `configure_*` helpers from the 3300-line `src/layer1/building.rs` monolith.

The file was converted into a facade (`src/layer1/building/mod.rs`), and the extracted logic was moved into a highly cohesive `src/layer1/building/configuration.rs` submodule. Test imports were updated to ensure clean compilation without triggering clippy unused warnings.

+++ src/experimental/psychic_resonance/mod.rs
@@ -41,10 +41,10 @@ pub fn psychic_resonance_system(
                 // High value items give positive resonance
                 match item.resource_type {
                     ResourceType::Alcohol | ResourceType::Rations | ResourceType::Metal => {
-                        resonance_delta += RESONANCE_STRENGTH * (item.amount as f32).min(10.0);
+                        resonance_delta += RESONANCE_STRENGTH * item.amount.min(10.0);
                     }
                     ResourceType::Waste | ResourceType::Scrap => {
-                        resonance_delta -= RESONANCE_STRENGTH * (item.amount as f32).min(10.0);
+                        resonance_delta -= RESONANCE_STRENGTH * item.amount.min(10.0);
                     }
                     _ => {}
                 }
